project/code/requirements.txt
keras
tensorflow
pandas
sklearn
numpy
matplotlib
cloudmesh-common
cloudmesh-cmd5
jupyter
yfinance
pathlib